requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0
alembic>=1.13
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9
httpx>=0.27
openai>=1.30
Pillow>=10.0
PyPDF2>=3.0
python-dotenv>=1.0
boto3>=1.34
psycopg2-binary>=2.9
aiofiles>=24.1
